Combine remaining new standup steps into single component
